@@ -20,21 +20,23 @@
     "url": "https://github.com/tclindner/npm-package-json-lint.git"
   },
   "bin": {
-    "npmPkgJsonLint": "src/cli.js"
+    "npmPkgJsonLint": "dist/cli.js"
   },
   "files": [
-    "src",
     "CONTRIBUTING.md",
-    "index.d.ts"
+    "dist"
   ],
-  "main": "src/api.js",
-  "types": "index.d.ts",
+  "main": "dist/api.js",
+  "types": "dist/src/api.d.ts",
   "scripts": {
+    "build": "npm run esbuild && npm run tsc",
+    "esbuild": "node esbuild.config.js",
     "eslint": "eslint . --format=node_modules/eslint-formatter-pretty",
-    "npmpackagejsonlint": "node src/cli.js ./package.json",
+    "npmpackagejsonlint": "node dist/cli.js ./package.json",
     "lint": "npm run eslint && npm run npmpackagejsonlint",
     "test": "jest",
-    "test:ci": "jest --runInBand"
+    "test:ci": "jest --runInBand",
+    "tsc": "tsc --project tsconfig.json"
   },
   "dependencies": {
     "ajv": "^8.10.0",
@@ -42,35 +44,44 @@
     "chalk": "^4.1.2",
     "cosmiconfig": "^7.0.1",
     "debug": "^4.3.3",
-    "globby": "^11.0.4",
+    "globby": "^11.1.0",
     "ignore": "^5.2.0",
     "is-plain-obj": "^3.0.0",
     "jsonc-parser": "^3.0.0",
     "log-symbols": "^4.1.0",
-    "meow": "^6.1.1",
+    "meow": "^9.0.0",
     "plur": "^4.0.0",
     "semver": "^7.3.5",
     "slash": "^3.0.0",
     "strip-json-comments": "^3.1.1",
+    "type-fest": "^2.12.0",
     "validate-npm-package-name": "^3.0.0"
   },
   "devDependencies": {
-    "eslint": "^7.32.0",
-    "eslint-config-tc": "^20.0.0",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.21",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "esbuild": "^0.14.23",
+    "esbuild-node-externals": "^1.4.1",
+    "eslint": "^8.10.0",
+    "eslint-config-tc": "^21.0.0",
+    "eslint-config-typescript-tc": "^5.0.0",
     "eslint-formatter-pretty": "^4.1.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jest": "^25.7.0",
+    "eslint-plugin-jest": "^26.1.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-unicorn": "^40.1.0",
+    "eslint-plugin-unicorn": "^41.0.0",
     "figures": "^3.2.0",
     "jest": "^27.5.1",
     "npm-package-json-lint-config-default": "^3.0.0",
     "npm-package-json-lint-config-tc": "^4.1.0",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "ts-jest": "^27.1.3",
+    "typescript": "^4.5.5"
   },
   "engines": {
-    "node": ">=10.0.0",
+    "node": ">=14.0.0",
     "npm": ">=6.0.0"
   },
   "license": "MIT"
